Fix cloning new repos/plugins
When updating a plugin, the script enters each plugins directory
but never exits it. Plugins became nested and could never be
properly updated or installed. By stepping back down to the root
directory after updating, this problem can be fixed.

In order to diagnose this issue, the quiet flag was removed from the
git command. Since the noise caused by git was small to begin with,
it makes sense just to leave it this way.